Remove network-manager installation
This install seems to mess up the networking on the Debian instance.
It is only being installed to placate the compliance scan, since that
scan uses nmcli to check that all wireless interfaces are turned off.
Since there are no wireless interfaces on the AWS instances, we can
state that the failed check is a false positive.