Ensure that Ansible intreprets the permission octal number as a string
Before I used 0600, which Ansible interpreted as an octal number and
changed to 384.

Also remove the enveloping block from the tasks in
fix_grub_config_perms.  It would be great to put that whole mess in a
block and let that block's become: yes and tags: grub apply to all
these tasks.  Unfortunately that does not appear to work.  When I try
that I get error messages from Ansible that look like "Unexpected
exception: Block object has no attribute get_name".

Finally, give the handler that fixes the grub config file permissions
a unique name.  I don't know that this is necessary, but I did it
while debugging the issue described in the previous paragraph.  It
doesn't hurt.